golang study >> ginchat message send recv

+++ ginchat/models/contact.go
@@ -4,8 +4,8 @@ import "gorm.io/gorm"
 
 type Contact struct {
 	gorm.Model
-	OwnerId  uint
-	TargetId uint
+	OwnerId  int64
+	TargetId int64
 	Type     int
 }
